Update WebSocket Server to handle multiple connections.
This pull request is incomplete (due to my inexperience in js...), but hoping someone else can help finish it off. The code change works to handle to multiple connections, what is missing is removing a disconnected session from the connections array. In addition, I am not sure why, but the sendData function still tries to send data to a closed connection and logs an error.